Refactor out the `add` methods
